+++ fm2prof/Fm2ProfRunner.py
@@ -18,7 +18,7 @@ from fm2prof.common import FM2ProfBase
 from fm2prof.CrossSection import CrossSection, CrossSectionHelpers
 from fm2prof.Export import Export1DModelData, OutputFiles
 from fm2prof.Import import FMDataImporter, FmModelData, ImportInputFiles
-from fm2prof.IniFile import IniFile
+from fm2prof.ini_file import IniFile
 from fm2prof.MaskOutputFile import MaskOutputFile
 from fm2prof.RegionPolygonFile import RegionPolygonFile, SectionPolygonFile
