+++ fm2prof/cli.py
@@ -3,8 +3,9 @@ from typing import Optional
 
 import typer
 from tqdm import tqdm
+
 from fm2prof import Project, __version__
-from fm2prof.IniFile import IniFile
+from fm2prof.ini_file import IniFile
 from fm2prof.utils import Compare1D2D, VisualiseOutput
 
 app = typer.Typer()
@@ -38,10 +39,9 @@ def cli_check_project(projectname: str) -> None:
 
 @app.command("compare")
 def cli_compare_1d2d(
-    projectname: str, output_1d: str, output_2d: str, routes: str
+    projectname: str, output_1d: str, output_2d: str, routes: str,
 ) -> None:
     """BETA FUNCTIONALITY - compares 1D and 2D results"""
-
     cf = Path(projectname).with_suffix(".ini")
     project = Project(cf)
 
@@ -49,7 +49,7 @@ def cli_compare_1d2d(
     path_2d = Path(output_2d)
 
     plotter = Compare1D2D(
-        project=project, path_1d=path_1d, path_2d=path_2d, routes=routes
+        project=project, path_1d=path_1d, path_2d=path_2d, routes=routes,
     )
 
     plotter.eval()
@@ -59,7 +59,7 @@ def cli_compare_1d2d(
 def cli_load_project(
     projectname: str,
     overwrite: bool = typer.Option(
-        False, "--overwrite", "-o", help="Overwrite if output already exists"
+        False, "--overwrite", "-o", help="Overwrite if output already exists",
     ),
     pp: bool = typer.Option(
         False,
@@ -75,7 +75,7 @@ def cli_load_project(
 
     if pp:
         vis = VisualiseOutput(
-            project.get_output_directory(), logger=project.get_logger()
+            project.get_output_directory(), logger=project.get_logger(),
         )
         for css in tqdm(vis.cross_sections):
             vis.figure_cross_section(css)
@@ -95,7 +95,6 @@ def cli(
     ),
 ) -> None:
     typer.echo("Welcome to Fm2Prof")
-    return
 
 
 def main():
